chore: add EVAL_CANARY_INTERVAL and EVAL_CANARY_LOOKBACK_MINUTES to .env
