+++ .env
@@ -59,6 +59,10 @@ NETWORK_NAME=observability-stack-network
 WEATHER_AGENT_PORT=8000
 CANARY_INTERVAL=120
 
+# Agent Eval Canary Configuration
+EVAL_CANARY_INTERVAL=120
+EVAL_CANARY_LOOKBACK_MINUTES=15
+
 # ******************
 # OpenTelemetry Demo Configuration
 # ******************
