+++ docker-compose/agent-eval-canary/Dockerfile
